Boost: Fix skip proxy origins to load css during critical CSS generation (#20793)
* [not verified] Add filter to override critical css proxy skipped hosts

While generating critical CSS, the CSS files need to be proxied through a the plugin to avoid CORS restrictions. Previously the proxy was skipped if the server hostname matched the request. Or if there was no hostname provided in the url. Now, it has a filter hook to allow more control.

* [not verified] Add changelogger

* [not verified] Replace proxy skipped host filter with fixed origin check

The host filter was implemented to overcome the git development issue. But the real problem was that, the hostname was not considering port while comparing.

* Update changelogger

* Fix foce css proxy condition

* Add test case for `force_proxied_css()`

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1181131867