Reconfigure eslint to pull from `.gitignore` and per-dir `.eslintignore` (#22138)
* Reconfigure eslint to pull from `.gitignore` and per-dir `.eslintignore`

Unfortunately we have to do it in a somewhat hacky manner: in eslintrc
we run a function that reads all the ignore files and dumps the results
into eslint's `ignorePatterns` config.

Then we have to do it all over again for any sub-eslintrc that does
`root: true` or that extends any eslintrc that has the hack. Inheritance
is ok, but apparently we don't do that all that often (sigh).

* Move ignore rule from merge into projects/packages/search/.eslintignore

* Add debug dep in tools/js-tools

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1723904578