Widget Visibility: switch to using shared js package utility (#22741)
In #22527, I copied over a utility function that we used in the block bundle to use in the Widget visibility bundle. However, starting in #22733, that utility function is now part of an external js package that we can use in the Widget visiblity bundle. Let's do that and save some duplication!

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1814545855