jetpack: Include a WAF composer dep in the build (#23094)
PR #22816 added packages/waf to Jetpack, and that package has a
(non-dev) Composer dep. If we want the package to work (someday, looks
like there's no UI for enabling or configuring it yet), that dep will
need to be included in the plugin bundle.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1899101106