linting: Check that packages with i18n set `.type` to `jetpack-library` (#23604)
In order for the i18n to actually work, a package with i18n strings must
set its `type` in composer.json to "jetpack-library" so
jetpack-composer-plugin will put it into `jetpack_vendor/` instead of
`vendor/`.

This is easy to overlook, so let's add a check for it.

Then fix existing packages that need it.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2038633760