boost: Remove outdated rollup polyfill plugin (#24093)
`rollup-plugin-node-builtins` hasn't been updated in 5 years, and
depends on really old versions of a lot of packages.

It turns out that removing this plugin makes no difference to the build:
it still succeeds, and the built files are identical. So let's do that.
If at some point in the future it turns out that a polyfill is needed
again, add one that's up to date (currently `rollup-plugin-polyfill-node`).

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2228736697