My Jetpack: Add VideoPress stats selector (#29764)
* Add actions to set product stats and fetching state

* Add reducer to update state with product stats and fetching status

* Add selector to read product stats and fetching status data from the state

* Add changelog file

* Rename changelog file to match branch name

* Add resolver for the product stats

Given each product has it's own endpoint and it's own set of relevant stats, the resolver will delegate the resolution of some product stats to a product-specific resolver.

* Update VideoPress stats resolver to get data from the new featured stats endpoint

* Set stats to null when the request fails

This way, the requester can know that it's not available and can rely on the null value.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4567104619