tools: Update output for `check-intra-monorepo-deps.sh -R` (#31408)
When cherry-picking package changes to a relase branch, one of the early
steps is to run `changelogger-release.sh -R`, which calls
`check-intra-monorepo-deps.sh -R`, which prints a reminder that the
dependencies are being incompletly updated and what the next steps are.

We're updating those steps now, so the reminder also needs updating.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5282921344