Activity Log: add link to Activity Log to Jetpack menu (#34174)
* Activity Log: add link to Activity Log to Jetpack menu

The Activity Log is a feature that can be very useful for site owners, but that we typically haven't shown a lot until now. This commit has an "Activity Log" menu item to the Jetpack menu, for all Jetpack plugins using "My Jetpack", as soon as the user is connected to WordPress.com.

Just like for the "My Jetpack" menu, we do not show it on Multisite sites since Jetpack Cloud isn't fully compatible with Multisite yet.

* Adjust comment

* Fix the focus when in the Jetpack dashboard

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/6899839140