image-cdn: Avoid fatal on bad img width/height (#39208)
If the `<img>` tag has a value for `width` or `height` that's neither an
integer nor a percentage, this will cause a fatal error when the value
is attempted to be used as an integer. Add validation to avoid this.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/10686704370

Upstream-Ref: Automattic/jetpack@2d7ee3082f6d866ac396a15cb04f7f6f83fbb162